+++ core/src/main/java/com/vanpra/composematerialdialogs/MaterialDialogLists.kt
@@ -20,6 +20,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.onCommit
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -102,7 +103,8 @@ fun <T> MaterialDialog.listItems(
             content = {
                 list.forEachIndexed { index, it ->
                     Box(
-                        Modifier.fillMaxWidth()
+                        Modifier
+                            .fillMaxWidth()
                             .clickable(
                                 onClick = {
                                     if (closeOnClick) {
@@ -155,7 +157,8 @@ fun MaterialDialog.listItemsMultiChoice(
             }
         }
     }
-    remember {
+
+    onCommit {
         if (waitForPositiveButton) {
             callbacks.add {
                 onCheckedChange(selectedItems)
@@ -175,11 +178,17 @@ fun MaterialDialog.listItemsMultiChoice(
         val selected = index in selectedItems
 
         Row(
-            Modifier.fillMaxWidth().preferredHeight(48.dp),
+            Modifier
+                .fillMaxWidth()
+                .preferredHeight(48.dp),
             verticalAlignment = Alignment.CenterVertically
         ) {
             Checkbox(checked = selected, onCheckedChange = { onChecked(index) }, enabled = enabled)
-            Spacer(modifier = Modifier.fillMaxHeight().width(32.dp))
+            Spacer(
+                modifier = Modifier
+                    .fillMaxHeight()
+                    .width(32.dp)
+            )
             Text(
                 item,
                 color = if (enabled) {
@@ -212,7 +221,7 @@ fun MaterialDialog.listItemsSingleChoice(
     onChoiceChange: (selected: Int) -> Unit = {}
 ) {
     val disableIndex = remember { positiveEnabled.size }
-    remember {
+    onCommit {
         positiveEnabled.add(disableIndex, initialSelection != null)
     }
 
@@ -232,7 +241,7 @@ fun MaterialDialog.listItemsSingleChoice(
         }
     }
 
-    remember {
+    onCommit {
         if (waitForPositiveButton) {
             callbacks.add { onChoiceChange(selected!!) }
         }
@@ -270,7 +279,9 @@ private fun SingleChoiceItem(
     val enabled = remember(disabledIndices) { index !in disabledIndices }
 
     Row(
-        Modifier.fillMaxWidth().preferredHeight(48.dp),
+        Modifier
+            .fillMaxWidth()
+            .preferredHeight(48.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
         RadioButton(
@@ -282,7 +293,11 @@ private fun SingleChoiceItem(
             },
             enabled = isEnabled(index)
         )
-        Spacer(modifier = Modifier.fillMaxHeight().width(32.dp))
+        Spacer(
+            modifier = Modifier
+                .fillMaxHeight()
+                .width(32.dp)
+        )
         Text(
             item,
             modifier = Modifier,